+++ packages/contracts-bedrock/contracts/test/L2OutputOracle.t.sol
@@ -10,17 +10,6 @@ import { Types } from "../libraries/Types.sol";
 contract L2OutputOracleTest is L2OutputOracle_Initializer {
     bytes32 proposedOutput1 = keccak256(abi.encode(1));
 
-    event OutputProposed(
-        bytes32 indexed outputRoot,
-        uint256 indexed l2OutputIndex,
-        uint256 indexed l2BlockNumber,
-        uint256 l1Timestamp
-    );
-
-    function setUp() public override {
-        super.setUp();
-    }
-
     function test_constructor_succeeds() external {
         assertEq(oracle.PROPOSER(), proposer);
         assertEq(oracle.CHALLENGER(), owner);
@@ -39,6 +28,40 @@ contract L2OutputOracleTest is L2OutputOracle_Initializer {
             startingBlockNumber,
             // startingTimestamp is in the future
             block.timestamp + 1,
+            proposer,
+            owner
+        );
+    }
+
+    function test_constructor_l2BlockTimeZero_reverts() external {
+        vm.expectRevert("L2OutputOracle: L2 block time must be greater than 0");
+        new L2OutputOracle(
+            submissionInterval,
+            0,
+            startingBlockNumber,
+            block.timestamp,
+            proposer,
+            owner
+        );
+    }
+
+    function testFuzz_constructor_submissionIntervalLteL2BlockTime_reverts(
+        uint256 _submissionInterval,
+        uint256 _l2BlockTime
+    ) external {
+        // Bound the _l2blockTime to be in the range of [1, type(uint256).max]
+        _l2BlockTime = bound(_l2BlockTime, 1, type(uint256).max);
+        // Roll the block number to _l2blockTime (the starting L2 timestamp must be less than or equal to the current time)
+        vm.roll(_l2BlockTime);
+        // Bound _submissionInterval to be less than or equal to _l2BlockTime
+        _submissionInterval = bound(_submissionInterval, 0, _l2BlockTime);
+
+        vm.expectRevert("L2OutputOracle: submission interval must be greater than L2 block time");
+        new L2OutputOracle(
+            _submissionInterval,
+            _l2BlockTime,
+            startingBlockNumber,
+            block.timestamp,
             proposer,
             owner
         );
@@ -300,8 +323,6 @@ contract L2OutputOracleTest is L2OutputOracle_Initializer {
     /*****************************
      * Delete Tests - Happy Path *
      *****************************/
-
-    event OutputsDeleted(uint256 indexed prevNextOutputIndex, uint256 indexed newNextOutputIndex);
 
     function test_deleteOutputs_singleOutput_succeeds() external {
         test_proposeL2Output_proposeAnotherOutput_succeeds();